# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pangolin Catalog API
  description: Multi-tenant Apache Iceberg REST Catalog with advanced features including federated catalogs, credential vending, and fine-grained permissions
  contact:
    name: Pangolin API Support
    email: support@pangolin.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
  /api/v1/access-requests:
    get:
      tags:
      - Business Metadata
      operationId: list_access_requests
      responses:
        '200':
          description: List of access requests
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/access-requests/{request_id}:
    get:
      tags:
      - Business Metadata
      operationId: get_access_request
      parameters:
      - name: request_id
        in: path
        description: Request ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Request details
        '403':
          description: Forbidden
        '404':
          description: Request not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    put:
      tags:
      - Business Metadata
      operationId: update_access_request
      parameters:
      - name: request_id
        in: path
        description: Request ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestStatus'
        required: true
      responses:
        '200':
          description: Request updated
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Request not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/assets/search:
    get:
      tags:
      - Business Metadata
      operationId: search_assets
      responses:
        '200':
          description: Search results
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/assets/{asset_id}:
    get:
      tags:
      - Business Metadata
      operationId: get_asset_details
      parameters:
      - name: asset_id
        in: path
        description: Asset ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Asset details with metadata
        '403':
          description: Forbidden
        '404':
          description: Asset not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/assets/{asset_id}/access-requests:
    post:
      tags:
      - Business Metadata
      operationId: request_access
      parameters:
      - name: asset_id
        in: path
        description: Asset ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessRequestPayload'
        required: true
      responses:
        '201':
          description: Access request created
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/assets/{asset_id}/metadata:
    get:
      tags:
      - Business Metadata
      operationId: get_business_metadata
      parameters:
      - name: asset_id
        in: path
        description: Asset ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Metadata retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataResponse'
        '404':
          description: Metadata not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Business Metadata
      operationId: add_business_metadata
      parameters:
      - name: asset_id
        in: path
        description: Asset ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMetadataRequest'
        required: true
      responses:
        '200':
          description: Metadata added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataResponse'
        '403':
          description: Forbidden
        '404':
          description: Asset not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    delete:
      tags:
      - Business Metadata
      operationId: delete_business_metadata
      parameters:
      - name: asset_id
        in: path
        description: Asset ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Metadata deleted
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
  /api/v1/catalogs:
    get:
      tags:
      - Catalogs
      operationId: list_catalogs
      responses:
        '200':
          description: List of catalogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogResponse'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Catalogs
      operationId: create_catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCatalogRequest'
        required: true
      responses:
        '201':
          description: Catalog created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        '400':
          description: Bad request
        '403':
          description: Forbidden for root user
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    AddMetadataRequest:
      type: object
      required:
      - tags
      - properties
      - discoverable
      properties:
        description:
          type: string
          nullable: true
        discoverable:
          type: boolean
        properties:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
...
    BusinessMetadata:
      type: object
      description: Business metadata attached to assets
      required:
      - id
      - asset-id
      - tags
      - properties
      - discoverable
      - created-by
      - created-at
      - updated-by
      - updated-at
      properties:
        asset-id:
          type: string
          format: uuid
        created-at:
          type: string
          format: date-time
        created-by:
          type: string
          format: uuid
        description:
          type: string
          description: Markdown description
          nullable: true
        discoverable:
          type: boolean
          description: Whether this asset appears in search for users without access
        id:
          type: string
          format: uuid
        properties:
          type: object
          description: Custom key-value properties
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
          description: Tags/labels for categorization
        updated-at:
          type: string
          format: date-time
        updated-by:
          type: string
          format: uuid
...
    CatalogResponse:
      type: object
      required:
      - id
      - name
      - catalog_type
      - properties
      properties:
        catalog_type:
          $ref: '#/components/schemas/pangolin_core.model.CatalogType'
        federated_config:
          allOf:
          - $ref: '#/components/schemas/pangolin_core.model.FederatedCatalogConfig'
          nullable: true
        id:
          type: string
          format: uuid
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
        storage_location:
          type: string
          nullable: true
        warehouse_name:
          type: string
          nullable: true
...
    CreateAccessRequestPayload:
      type: object
      properties:
        reason:
          type: string
          nullable: true
...
    CreateCatalogRequest:
      type: object
      required:
      - name
      properties:
        catalog_type:
          allOf:
          - $ref: '#/components/schemas/pangolin_core.model.CatalogType'
          nullable: true
        federated_config:
          allOf:
          - $ref: '#/components/schemas/pangolin_core.model.FederatedCatalogConfig'
          nullable: true
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
          nullable: true
        storage_location:
          type: string
          nullable: true
        warehouse_name:
          type: string
          nullable: true
...
    MetadataResponse:
      type: object
      required:
      - metadata
      properties:
        metadata:
          $ref: '#/components/schemas/BusinessMetadata'
...
    RequestStatus:
      type: string
      description: Status of an access request
      enum:
      - pending
      - approved
      - rejected
...
    UpdateRequestStatus:
      type: object
      required:
      - status
      properties:
        comment:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RequestStatus'
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/users/login or /api/v1/tokens
...